+++ src/components/call_out/index.d.ts
@@ -1,5 +1,5 @@
 import { CommonProps, Omit } from '../common';
-import { IconType } from '../icon'
+import { IconType } from '../icon';
 
 import { FunctionComponent, ReactNode, HTMLAttributes } from 'react';
 
@@ -14,13 +14,15 @@ declare module '@elastic/eui' {
   type Size = 's' | 'm';
 
   export interface EuiCallOutProps {
-    title?: ReactNode,
-    iconType?: IconType,
-    color?: Color,
-    size?: Size,
+    title?: ReactNode;
+    iconType?: IconType;
+    color?: Color;
+    size?: Size;
   }
 
   export const EuiCallOut: FunctionComponent<
-    CommonProps & EuiCallOutProps & Omit<HTMLAttributes<HTMLDivElement>, 'title'>
+    CommonProps &
+      EuiCallOutProps &
+      Omit<HTMLAttributes<HTMLDivElement>, 'title'>
   >;
 }