+++ src/components/key_pad_menu/key_pad_menu_item.test.js
@@ -42,7 +42,7 @@ describe('EuiKeyPadMenuItemButton', () => {
   });
 
   describe('onClick', () => {
-    test(`isn't called upon instantiation`, () => {
+    test("isn't called upon instantiation", () => {
       const onClickHandler = sinon.stub();
 
       shallow(