+++ src/components/selectable/selectable_list/selectable_list.test.tsx
@@ -91,7 +91,7 @@ describe('EuiSelectableListItem', () => {
           renderOption={(option: Option, searchValue?: string) => {
             return (
               <span>
-                {searchValue} => {option.label}
+                {searchValue} =&gt; {option.label}
               </span>
             );
           }}