+++ src/components/table/index.d.ts
@@ -4,6 +4,11 @@ import { CommonProps, NoArgCallback } from '../common';
 import { IconType } from '../icon';
 import { HorizontalAlignment } from '../../services/alignment';
 import { EuiTableRowCellCheckbox as TableRowCellCheckbox } from './table_row_cell_checkbox';
+import {
+  EuiTableHeaderCellCheckboxScope as TableHeaderCellCheckboxScope,
+  EuiTableHeaderCellCheckboxProps as TableHeaderCellCheckboxProps,
+  EuiTableHeaderCellCheckbox as TableHeaderCellCheckbox,
+} from './table_header_cell_checkbox';
 
 import {
   FunctionComponent,
@@ -12,7 +17,7 @@ import {
   ButtonHTMLAttributes,
   ThHTMLAttributes,
   TdHTMLAttributes,
-  ReactNode
+  ReactNode,
 } from 'react';
 
 declare module '@elastic/eui' {
@@ -49,7 +54,9 @@ declare module '@elastic/eui' {
 
   export interface EuiTableHeaderProps {}
 
-  export const EuiTableHeader: FunctionComponent<CommonProps & EuiTableHeaderProps>;
+  export const EuiTableHeader: FunctionComponent<
+    CommonProps & EuiTableHeaderProps
+  >;
 
   /**
    * table header button type defs
@@ -102,22 +109,9 @@ declare module '@elastic/eui' {
    * @see './table_header_cell_checkbox.js'
    */
 
-  export type EuiTableHeaderCellCheckboxScope =
-    | 'col'
-    | 'row'
-    | 'colgroup'
-    | 'rowgroup';
-
-  export interface EuiTableHeaderCellCheckboxProps {
-    width?: string;
-    scope?: EuiTableHeaderCellCheckboxScope;
-  }
-
-  export const EuiTableHeaderCellCheckbox: FunctionComponent<
-    CommonProps &
-      TdHTMLAttributes<HTMLTableCellElement> &
-      EuiTableHeaderCellCheckboxProps
-  >;
+  export type EuiTableHeaderCellCheckboxScope = TableHeaderCellCheckboxScope;
+  export interface EuiTableHeaderCellCheckboxProps extends TableHeaderCellCheckboxProps {}
+  export const EuiTableHeaderCellCheckbox: typeof TableHeaderCellCheckbox;
 
   /**
    * table row type defs
@@ -146,7 +140,7 @@ declare module '@elastic/eui' {
     truncateText?: boolean;
   }
 
-  export type EuiTableRowCellMobileOptionsShape = {
+  export interface EuiTableRowCellMobileOptionsShape {
     show?: boolean;
     only?: boolean;
     render?: ReactNode;
@@ -162,7 +156,8 @@ declare module '@elastic/eui' {
     isExpander?: boolean;
     isMobileFullWidth?: boolean;
     isMobileHeader?: boolean;
-    mobileOptions?: EuiTableRowCellMobileOptionsShape & EuiTableRowCellSharedPropsShape;
+    mobileOptions?: EuiTableRowCellMobileOptionsShape &
+      EuiTableRowCellSharedPropsShape;
   }
 
   export const EuiTableRowCell: FunctionComponent<